Split out the node and browser versions
