+++ package.json
@@ -31,6 +31,7 @@
     "np": "^6.5.0",
     "prettier-eslint": "^11.0.0",
     "rollup": "^2.28.2",
+    "serve": "^11.3.2"
   },
   "publishConfig": {
     "access": "public"
