+++ src/browser.js
@@ -1,0 +1,10 @@
+import GridIntensity from './gridintensity'
+
+GridIntensity.prototype.getLocalStoragePolyFill = function {
+  return localStorage
+}
+GridIntensity.prototype.getFetch = function () {
+  return fetch
+}
+
+export default { GridIntensity }
