+++ src/gridIntensity.js
@@ -1,0 +1,73 @@
+
+// import fetch from 'cross-fetch';
+import settings from './defaultSettings'
+
+const intensityProvider = settings.uk
+
+function GridIntensity() {
+  this.data = []
+  this.localStorage = this.getLocalStoragePolyFill()
+}
+
+
+GridIntensity.prototype.setup = async function (localStorage, fetch) {
+  let fetchNewData
+  this.data = this.getLocalIntensityData()
+
+  if (this.data.length < 1) {
+    this.data = await this.fetchIntensityData()
+  }
+}
+GridIntensity.prototype.getLocalIntensityData = function () {
+  //  if we're not in a browser, use a localstorage polyfill
+  let storage = this.getLocalStoragePolyFill()
+
+  // otherwise we're in a browser
+  const data = storage.getItem('gridIntensityData')
+  // we have no local data - return early
+  if (!data || data.length < 1) {
+    return []
+  }
+  console.log(data)
+  // try to parse what we already have
+  try {
+    parsed_data = JSON.parse(data)
+    console.log(parsed_data)
+    return parsed_data
+  } catch (err) {
+    // debug(`error, parsing the stored JSON`, err)
+    storage.setItem('gridIntensityData', [])
+    return []
+  }
+}
+
+GridIntensity.prototype.getCarbonIndex = async function (options) {
+  let now
+  if (options && options.checkDate) {
+    now = options.checkDate
+  } else {
+    now = new Date()
+  }
+
+  // this only fetches the last date. If we fetch more dates ahead, we need to
+  // find the most closest date in the set to now, as we'd have more than
+  // one to choose from
+  let latestReading = this.data.data[this.data.data.length - 1]
+  const latestReadingDate = Date.parse(latestReading.to)
+
+  if (now > latestReadingDate) {
+    // fetch new data, as this out of date
+    this.data = await this.fetchIntensityData()
+    latestReading = this.data.data[0]
+  }
+  return latestReading.intensity.index
+
+}
+
+GridIntensity.prototype.fetchIntensityData = async function () {
+  let res = await fetch(intensityProvider.api.current)
+  this.data = await res.json()
+  return this.data
+}
+
+export default GridIntensity
